refactor: move codegen logic to separate module
